+++ mne/fiff/tests/test_raw.py
@@ -1,4 +1,5 @@
 import os.path as op
+from copy import deepcopy
 
 import numpy as np
 from nose.tools import assert_true
@@ -67,6 +68,32 @@ def test_io_raw():
         fname = op.join(op.dirname(__file__), 'data', 'test_raw.fif')
 
 
+def test_io_complex():
+    """ Test IO with complex data types """
+    dtypes = [np.complex64, np.complex128]
+
+    raw = Raw(fif_fname, preload=True)
+    picks = np.arange(5)
+    start, stop = raw.time_to_index(0, 5)
+
+    data_orig, _ = raw[picks, start:stop]
+
+    for dtype in dtypes:
+        imag_rand = np.array(1j * np.random.randn(data_orig.shape[0],
+                            data_orig.shape[1]), dtype)
+
+        raw_cp = deepcopy(raw)
+        raw_cp._data = np.array(raw_cp._data, dtype)
+        raw_cp._data[picks, start:stop] += imag_rand
+        raw_cp.save('raw.fif', picks, tmin=0, tmax=5)
+
+        raw2 = Raw('raw.fif')
+        raw2_data, _ = raw2[picks, :]
+        n_samp = raw2_data.shape[1]
+        assert_array_almost_equal(raw2_data[:, :n_samp],
+                                  raw_cp._data[picks, :n_samp])
+
+
 def test_getitem():
     """Test getitem/indexing of Raw
     """
@@ -111,3 +138,48 @@ def test_preload_modify():
         data_new, _ = raw_new[picks, :nsamp / 2]
 
         assert_array_almost_equal(data, data_new)
+
+
+def test_filter():
+    """ Test filtering and Raw.apply_function interface """
+
+    raw = Raw(fif_fname, preload=True)
+    picks_meg = pick_types(raw.info, meg=True)
+    picks = picks_meg[:4]
+
+    raw_lp = deepcopy(raw)
+    raw_lp.low_pass_filter(picks, 4.0, verbose=0, n_jobs=2)
+
+    raw_hp = deepcopy(raw)
+    raw_hp.high_pass_filter(picks, 8.0, verbose=0)
+
+    raw_bp = deepcopy(raw)
+    raw_bp.band_pass_filter(picks, 4.0, 8.0, verbose=0)
+
+    data, _ = raw[picks, :]
+
+    lp_data, _ = raw_lp[picks, :]
+    hp_data, _ = raw_hp[picks, :]
+    bp_data, _ = raw_bp[picks, :]
+
+    assert_array_almost_equal(data, lp_data + hp_data + bp_data)
+
+    # make sure we didn't touch other channels
+    data, _ = raw[picks_meg[4:], :]
+    bp_data, _ = raw_bp[picks_meg[4:], :]
+
+    assert_array_equal(data, bp_data)
+
+
+def test_hilbert():
+    """ Test computation of analytic signal using hilbert """
+    raw = Raw(fif_fname, preload=True)
+    picks_meg = pick_types(raw.info, meg=True)
+    picks = picks_meg[:4]
+
+    raw2 = deepcopy(raw)
+    raw.apply_hilbert(picks, verbose=0)
+    raw2.apply_hilbert(picks, envelope=True, n_jobs=2, verbose=0)
+
+    env = np.abs(raw._data[picks, :])
+    assert_array_almost_equal(env, raw2._data[picks, :])