+++ doc/source/whats_new.rst
@@ -23,6 +23,9 @@ Changelog
 
    - Write BEM surfaces in Python  by `Alex Gramfort`_.
 
+   - Filtering operations and apply_function interface for Raw object by `Martin Luessi`_.
+
+   - Support for complex valued raw fiff files and computation of analytic signal for Raw object by `Martin Luessi`_.
 
 Version 0.2
 -----------